+++ CHANGELOG.md
@@ -12,7 +12,11 @@
 `
 #### BUG FIXES:
 
-* Fix bug in `akamai_cps_dv_enrollment` resource when MTLS settings are provided ([#339](https://github.com/akamai/terraform-provider-akamai/issues/339))
+* CPS
+  * Fix bug in `akamai_cps_dv_enrollment` resource when MTLS settings are provided ([#339](https://github.com/akamai/terraform-provider-akamai/issues/339))
+
+* GTM
+  * Make `test_object` inside `liveness_test` required only for `test_object_protocol` values: `HTTP`, `HTTPS` or `FTP` ([I#408](https://github.com/akamai/terraform-provider-akamai/issues/408))
 
 ## 4.1.0 (Jun 1, 2023)